Merge r145318 - Fix offset handling in GStreamer WebKitWebSource. https://bugs.webkit.org/show_bug.cgi?id=111888
Patch by Sebastian Dröge <sebastian.droege@collabora.co.uk> on 2013-03-09
Reviewed by Philippe Normand.

* platform/graphics/gstreamer/WebKitWebSourceGStreamer.cpp:
(webKitWebSrcSeekDataCb):
(StreamingClient::didReceiveData):
The offset handling in WebKitWebSource was completely wrong
before and caused wrong offsets to be set on the GStreamer buffers
after a seek.

Apart from that there was also a race condition that happens
when a downstream element causes seeks in very short succession
and is switching between two different parts of the stream.

git-svn-id: https://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.0@145521 268f45cc-cd09-0410-ab3c-d52691b4dbfc